Created quiz_route blueprint and moved create quiz route to it
