+++ api/templates/account.html
@@ -22,7 +22,7 @@
 
         <div class="action-block">
           <h2>Create a New Quiz</h2>
-          <a href="{{ url_for('create') }}" title="Create a quiz" class="button">Create</a>
+          <a href="{{ url_for('quiz.create') }}" title="Create a quiz" class="button">Create</a>
         </div>
         
         <div class="action-block">
