+++ api/templates/create.html
@@ -21,7 +21,7 @@
     {% endwith %}
 
     {% if not data %}
-    <form action="{{ url_for('create') }}" method="POST" id="createQuizForm">
+    <form action="{{ url_for('quiz.create') }}" method="POST" id="createQuizForm">
       <div class="form-group">
         <label>Quiz Title: <input type="text" name="title" required></label><br><br>
         <label>Description (300 characters max): <textarea name="description"></textarea></label><br><br>
@@ -54,7 +54,7 @@
 
     {% else %}
 
-    <form action="{{ url_for('create') }}" method="POST" id="createQuizForm">
+    <form action="{{ url_for('quiz.create') }}" method="POST" id="createQuizForm">
 
       <div class="form-group">
         <label>Quiz Title: <input type="text" name="title" value="{{ data.title }}" required></label><br><br>
